Moves enums into enum subpackage
