+++ src/pyrite/enum/__init__.py
