+++ src/pyrite/types/__init__.py
@@ -4,7 +4,6 @@ from typing import Protocol, runtime_checkable, TYPE_CHECKING
 
 from .camera import CameraBase  # noqa: F401
 from .entity import Entity  # noqa: F401
-from .enums import RenderLayers, AnchorPoint, Layer, Anchor  # noqa: F401
 from .renderable import Renderable  # noqa: F401
 from .surface_sector import SurfaceSector  # noqa: F401
 from .static_decor import StaticDecor  # noqa: F401
