+++ tests/test_render_manager.py
@@ -13,7 +13,7 @@ from pygame.typing import Point
 
 sys.path.append(str(pathlib.Path.cwd()))
 
-from src.pyrite.types.enums import Layer, RenderLayers  # noqa:E402
+from src.pyrite.enum import Layer, RenderLayers  # noqa:E402
 from src.pyrite.core.renderer import (  # noqa:E402
     DefaultRenderManager,
     _get_draw_index,
